+++ Encrypt.cs
@@ -3,7 +3,7 @@ using System.Numerics;
 
 public static class Encrypt
 {
-    public static int encryptMessage(BigInteger n, BigInteger e, int m)
+    public static int encryptMessage(BigInteger n, BigInteger e, byte m)
     {
         BigInteger c = BigInteger.ModPow(new BigInteger(m), e, n);
         return (int)c;
